src/app.controller.spec.ts
import { Test, TestingModule } from '@nestjs/testing';
import { AppController } from './app.controller';

describe('AppController', () => {
  let appController: AppController;

  beforeEach(async () => {
    const app: TestingModule = await Test.createTestingModule({
      controllers: [AppController],
      // If your controller relies on services or providers, they need to be included here
      // providers: [AppService],
    }).compile();

    appController = app.get<AppController>(AppController);
  });

  describe('findAll', () => {
    it('should return an array of entities', () => {
      expect(appController.findAll()).toBeInstanceOf(Array);
    });
  });

  describe('findOne', () => {
    it('should return a single entity', () => {
      // Mocking a simple find one functionality
      // In real scenarios, you might mock the service and its response
      const result = appController.findOne('1');
      expect(result).toBeDefined();
    });
  });

  describe('create', () => {
    it('should return a success message', () => {
      expect(appController.create({ name: 'New Entity' })).toBe('Entity created');
    });
  });

  describe('update', () => {
    it('should return an update message', () => {
      expect(appController.update('1', { name: 'Updated Entity' })).toBe('Entity updated');
    });
  });

  describe('delete', () => {
    it('should return a delete message', () => {
      expect(appController.delete('1')).toBe('Entity removed');
    });
  });
});
